Makefile automated push

+++ libft/srcs/str/ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: ldevelle <ldevelle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 17:11:36 by ldevelle          #+#    #+#             */
-/*   Updated: 2018/11/13 17:51:55 by ldevelle         ###   ########.fr       */
+/*   Updated: 2018/11/20 19:40:12 by ldevelle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strncat(char *dest, const char *src, size_t n)
 {
-	int	size;
-	int	i;
+	size_t	size;
+	size_t	i;
 
 	i = -1;
 	size = ft_strlen(dest);
