+++ libft/srcs/str/ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: ldevelle <ldevelle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 17:11:14 by ldevelle          #+#    #+#             */
-/*   Updated: 2018/11/13 17:31:22 by ldevelle         ###   ########.fr       */
+/*   Updated: 2018/11/20 19:39:53 by ldevelle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dest, const char *src, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = -1;
 	while (++i < n)
