+++ libft/srcs/str/ft_strnequ.c
@@ -6,7 +6,7 @@
 /*   By: ldevelle <ldevelle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 11:49:47 by ldevelle          #+#    #+#             */
-/*   Updated: 2018/11/14 11:51:18 by ldevelle         ###   ########.fr       */
+/*   Updated: 2018/11/20 19:39:14 by ldevelle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strnequ(char const *s1, char const *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] == s2[i] && i <= n)
