+++ libft/srcs/str/ft_strsub.c
@@ -6,7 +6,7 @@
 /*   By: ldevelle <ldevelle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/12 11:54:23 by ldevelle          #+#    #+#             */
-/*   Updated: 2018/11/16 22:16:01 by ldevelle         ###   ########.fr       */
+/*   Updated: 2018/11/20 19:32:06 by ldevelle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,15 +14,15 @@
 
 char	*ft_strsub(char const *s, unsigned int start, size_t len)
 {
-	int		i;
-	int		size;
-	char	*str;
+	size_t			i;
+	unsigned int	size;
+	char			*str;
 
 	i = -1;
 	size = ft_strlen(s);
 	if (size < start)
 		return (NULL);
-	if (!(str = (char*)ft_strnew(len)))
+	if (!(str = (char*)ft_strnew((int)len)))
 		return (NULL);
 	while (++i < len)
 	{
